move helper functions to helper file
